requirements.txt
PyQt5==5.15.7
attrs==22.1.0
matplotlib==3.6.2
more_itertools==9.0.0
numpy==1.23.0
sounddevice==0.4.5
soundfile==0.11.0